Fix the structure for geocode

+++ trunk/petsitter/src/main/java/com/sjsu/petsitter/bean/PetOwnerResult.java
@@ -28,8 +28,8 @@ public class PetOwnerResult {
     private String homePhone;
     private String mobile;
     private Double averageRating;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     	
     public String getDisplayName() {
         return displayName;
@@ -135,19 +135,19 @@ public class PetOwnerResult {
         this.averageRating = averageRating;
     }
     
-    public BigDecimal getLatitude() {
+    public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(BigDecimal latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public BigDecimal getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(BigDecimal longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
